Final Assignment Version (pre bonus)
I adjusted the height of .heroImg and added margin to the top to play with the space between that section and the footer. Additionally, I added my name in the header - some adjusting needs to happen here with spacing but it's at an okay point for now. I also added a white version of my logo and a background in the header to add more structure.